+++ aliyun-python-sdk-core/alibabacloud/__init__.py
@@ -1,0 +1,9 @@
+import logging
+
+
+class NullHandler(logging.Handler):
+    def emit(self, record):
+        pass
+
+
+logging.getLogger('alibabacloud').addHandler(NullHandler())
