+++ aliyun-python-sdk-core/aliyunsdkcore/__init__.py
@@ -1,5 +1,3 @@
-__version__ = "2.13.4"
-
 import logging
